Fixed quantized indexes not ranking as quickly as they should.
The introduction of the pregens with a way to mix them in with the ranker
made all the ranking use doubles to rank. This completely negated the point
of using shorts and quantized indexes.

If you want to use pregens now, define USE_PREGEN, which will slow the
ranking down.